--- Models/AppDbContext.cs
@@ -1,128 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using Microsoft.EntityFrameworkCore;
-
-namespace SmartSales_Api.Models;
-
-public partial class AppDbContext : DbContext
-{
-    public AppDbContext()
-    {
-    }
-
-    public AppDbContext(DbContextOptions<AppDbContext> options)
-        : base(options)
-    {
-    }
-
-    public virtual DbSet<Categoria> Categorias { get; set; }
-
-    public virtual DbSet<Producto> Productos { get; set; }
-
-    public virtual DbSet<Promocione> Promociones { get; set; }
-
-    public virtual DbSet<Role> Roles { get; set; }
-
-    public virtual DbSet<Usuario> Usuarios { get; set; }
-
-    protected override void OnConfiguring(DbContextOptionsBuilder optionsBuilder)
-#warning To protect potentially sensitive information in your connection string, you should move it out of source code. You can avoid scaffolding the connection string by using the Name= syntax to read it from configuration - see https://go.microsoft.com/fwlink/?linkid=2131148. For more guidance on storing connection strings, see https://go.microsoft.com/fwlink/?LinkId=723263.
-        => optionsBuilder.UseSqlServer("Server=(localdb)\\MSSQLLocalDB;Database=SmartSales;Trusted_Connection=True;MultipleActiveResultSets=true;TrustServerCertificate=True");
-
-    protected override void OnModelCreating(ModelBuilder modelBuilder)
-    {
-        modelBuilder.Entity<Categoria>(entity =>
-        {
-            entity.HasKey(e => e.Id).HasName("PK__Categori__3214EC072FB49D13");
-
-            entity.HasIndex(e => e.Nombre, "UQ__Categori__75E3EFCFF1B7927E").IsUnique();
-
-            entity.Property(e => e.Nombre).HasMaxLength(100);
-        });
-
-        modelBuilder.Entity<Producto>(entity =>
-        {
-            entity.HasKey(e => e.Id).HasName("PK__Producto__3214EC07F678B235");
-
-            entity.Property(e => e.CategoriaId).HasColumnName("Categoria_Id");
-            entity.Property(e => e.FechaCreacion)
-                .HasDefaultValueSql("(sysdatetime())")
-                .HasColumnName("Fecha_Creacion");
-            entity.Property(e => e.ImagenUrl).HasColumnName("Imagen_Url");
-            entity.Property(e => e.Nombre).HasMaxLength(100);
-
-            entity.HasOne(d => d.Categoria).WithMany(p => p.Productos)
-                .HasForeignKey(d => d.CategoriaId)
-                .OnDelete(DeleteBehavior.ClientSetNull)
-                .HasConstraintName("FK_Productos_Categorias");
-        });
-
-        modelBuilder.Entity<Promocione>(entity =>
-        {
-            entity.HasKey(e => e.Id).HasName("PK__Promocio__3214EC07A58251AB");
-
-            entity.Property(e => e.FechaFin).HasColumnName("Fecha_Fin");
-            entity.Property(e => e.FechaInicio).HasColumnName("Fecha_Inicio");
-            entity.Property(e => e.FechaRegistro)
-                .HasDefaultValueSql("(sysdatetime())")
-                .HasColumnName("Fecha_Registro");
-            entity.Property(e => e.Fuente).HasMaxLength(100);
-            entity.Property(e => e.IdProducto).HasColumnName("Id_Producto");
-            entity.Property(e => e.IdUsuarioRegistro).HasColumnName("Id_Usuario_Registro");
-            entity.Property(e => e.PrecioOriginal)
-                .HasColumnType("numeric(10, 2)")
-                .HasColumnName("Precio_Original");
-            entity.Property(e => e.PrecioPromocion)
-                .HasColumnType("numeric(10, 2)")
-                .HasColumnName("Precio_Promocion");
-            entity.Property(e => e.UrlProducto)
-                .HasMaxLength(200)
-                .HasColumnName("Url_Producto");
-
-            entity.HasOne(d => d.IdProductoNavigation).WithMany(p => p.Promociones)
-                .HasForeignKey(d => d.IdProducto)
-                .OnDelete(DeleteBehavior.ClientSetNull)
-                .HasConstraintName("FK_Promociones_Productos");
-
-            entity.HasOne(d => d.IdUsuarioRegistroNavigation).WithMany(p => p.Promociones)
-                .HasForeignKey(d => d.IdUsuarioRegistro)
-                .OnDelete(DeleteBehavior.ClientSetNull)
-                .HasConstraintName("FK_Promociones_Usuarios");
-        });
-
-        modelBuilder.Entity<Role>(entity =>
-        {
-            entity.HasKey(e => e.Id).HasName("PK__Roles__3214EC07A462045A");
-
-            entity.HasIndex(e => e.Nombre, "UQ__Roles__75E3EFCFE6D92622").IsUnique();
-
-            entity.Property(e => e.Nombre).HasMaxLength(100);
-        });
-
-        modelBuilder.Entity<Usuario>(entity =>
-        {
-            entity.HasKey(e => e.Id).HasName("PK__Usuarios__3214EC07DCA1FF43");
-
-            entity.HasIndex(e => e.Email, "UQ__Usuarios__A9D10534E303E832").IsUnique();
-
-            entity.Property(e => e.Email).HasMaxLength(200);
-            entity.Property(e => e.FechaRegistro)
-                .HasDefaultValueSql("(sysdatetime())")
-                .HasColumnName("Fecha_Registro");
-            entity.Property(e => e.IdRol).HasColumnName("Id_Rol");
-            entity.Property(e => e.Nombre).HasMaxLength(100);
-            entity.Property(e => e.PasswordHash)
-                .HasMaxLength(200)
-                .HasColumnName("Password_Hash");
-
-            entity.HasOne(d => d.IdRolNavigation).WithMany(p => p.Usuarios)
-                .HasForeignKey(d => d.IdRol)
-                .OnDelete(DeleteBehavior.ClientSetNull)
-                .HasConstraintName("FK_Usuarios_Roles");
-        });
-
-        OnModelCreatingPartial(modelBuilder);
-    }
-
-    partial void OnModelCreatingPartial(ModelBuilder modelBuilder);
-}
