+++ SmartSales-Api.csproj
@@ -25,6 +25,7 @@
   <ItemGroup>
     <Folder Include="Interfaces\" />
     <Folder Include="Endpoints\" />
+    <Folder Include="Methods\" />
   </ItemGroup>
 
 </Project>
